refactor: view 관련 클래스명 변경
