+++ src/main/java/racingcar/Application.java
@@ -1,13 +1,13 @@
 package racingcar;
 
 import racingcar.controller.RaceController;
-import racingcar.view.InputView;
-import racingcar.view.OutputView;
+import racingcar.view.ConsoleInputView;
+import racingcar.view.ConsoleOutputView;
 
 public class Application {
     public static void main(String[] args) {
 
-        RaceController raceController = new RaceController(InputView.getInstance(), OutputView.getInstance());
+        RaceController raceController = new RaceController(ConsoleInputView.getInstance(), ConsoleOutputView.getInstance());
         raceController.play();
     }
 }
