--- src/main/java/racingcar/view/InputViewable.java
@@ -1,10 +1,0 @@
-package racingcar.view;
-
-public interface InputViewable {
-
-    String askCarName();
-
-    String askPlayCount();
-
-    String getUserInput();
-}
